Refactor to remove C from the Linux driver
